Move rollcontext to roll expression
