+++ d20/roll/__init__.py
@@ -5,10 +5,9 @@ import random
 from collections.abc import Mapping
 from typing import Callable, Optional, Type
 
-from .expression import BinOp, Dice, Expression, Literal, Number, Parenthetical, UnOp
+from .expression import BinOp, Dice, Expression, Literal, Number, Parenthetical, UnOp, RollContext
 from .stringifier import SimpleStringifier, Stringifier
 from .. import diceast as ast, utils
-from ..context import RollContext
 from ..enums import AdvType, CritType
 from ..rand import random_impl
